Move util::format implementation to source file
- add format.cpp as source to benchmarks that don't link to HPX
