+++ tests/performance/local/CMakeLists.txt
@@ -169,6 +169,15 @@ foreach(benchmark ${benchmarks})
                      HPX_PREFIX ${HPX_BUILD_PREFIX}
                      FOLDER "Benchmarks/Local")
 
+  list(FIND ${benchmark}_FLAGS NOLIBS has_NOLIBS)
+  if (NOT has_NOLIBS EQUAL -1)
+    target_compile_definitions(${benchmark}_test PRIVATE
+                               HPX_FORMAT_EXPORT= #empty
+                               HPX_NO_VERSION_CHECK)
+    target_sources(${benchmark}_test
+                   PRIVATE ${PROJECT_SOURCE_DIR}/src/util/format.cpp)
+  endif()
+
   if(NOT "${${benchmark}_PARAMETERS}" MATCHES NO_HPX_MAIN)
     add_hpx_test("tests.performance.local" ${benchmark} ${${benchmark}_PARAMETERS})
   endif()
