+++ tests/performance/local/delay_baseline_threaded.cpp
@@ -3,8 +3,6 @@
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
-#define HPX_NO_VERSION_CHECK
-
 #include "worker_timed.hpp"
 
 #include <hpx/compat/barrier.hpp>
