+++ tests/performance/local/htts_v2/CMakeLists.txt
@@ -47,6 +47,15 @@ foreach(benchmark ${benchmarks})
                      HPX_PREFIX ${HPX_BUILD_PREFIX}
                      FOLDER "Benchmarks/HTTS v2/${benchmark}")
 
+  list(FIND ${benchmark}_FLAGS NOLIBS has_NOLIBS)
+  if (NOT has_NOLIBS EQUAL -1)
+    target_compile_definitions(${benchmark} PRIVATE
+                               HPX_FORMAT_EXPORT= #empty
+                               HPX_NO_VERSION_CHECK)
+    target_sources(${benchmark}
+                   PRIVATE ${PROJECT_SOURCE_DIR}/src/util/format.cpp)
+  endif()
+
   # add a custom target for this example
   add_hpx_pseudo_target(tests.performance.local.htts_v2.${benchmark})
