+++ tests/performance/local/htts_v2/htts2.cpp
@@ -5,7 +5,7 @@
 //
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
-#define HPX_NO_VERSION_CHECK
+
 #include "htts2.hpp"
 
 #include <boost/program_options.hpp>
