+++ tests/performance/local/htts_v2/htts2_payload_precision.cpp
@@ -5,7 +5,6 @@
 //
 //  Distributed under the Boost Software License, Version 1.0. (See accompanying
 //  file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
-#define HPX_NO_VERSION_CHECK
 
 #include "htts2.hpp"
 #include <hpx/util/format.hpp>
