+++ tests/performance/local/openmp_homogeneous_timed_task_spawn.cpp
@@ -29,8 +29,6 @@
 // NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
 // EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-#define HPX_NO_VERSION_CHECK
-
 #include <hpx/config.hpp>
 
 #include <hpx/util/assert.hpp>
